CRAYSAT-1314: Add product catalog query code
Add a new module, `query`, for clients to query the product catalog for
information about the installed product versions along with information
about the components provided by those products.

Test Description:
Tested by installing into a Python virtualenv on baldar and running a
script which used the `ProductCatalog` class to print information about
installed versions of products.

Tested by deploying a job in K8s and running a simple python command to
get the installed versions of the SAT product. This was just to ensure
that the `load_k8s` function worked to load the K8s config while running
inside a K8s pod.